feat(build): add build system (#2)
We'll build python packages (eventually) by running standard wheel
builds in a correct environment, provided by a buildroot SDK. The
buildroot SDK's environment will overwrite the current environment
because it gets sourced. In interactive contexts, you can just close
that shell when you're done, but we need to isolate further. Therefore,
we need a docker container.

This docker container provides a python environment so we can run python
build orchestration, but its big job is to
- Download the buildroot SDK (from our S3, where it is now a build
output of our buildroot)
- Contain the environment variable stuff that the SDK does within it

With this orchestration, we can do that.

We'll provide a python package of our own called `builder` (bad name if
it was published, but it's not) that has two entrypoints: a host entry
point that can be called to initiate a build and is vaguely
cross-python-version compatible, and a container entrypoint that runs
inside the container to execute the build. The host entrypoint builds
and runs the container (and is called from a quick script at the top of
the repo), and the container entrypoint for now activates the buildroot
environment and builds a test c file to prove that it gets the right
architecture.

Closes RCORE-208