feat(build): build wheels (#3)
Add the code harnessing to properly build wheels for the OT-2.

This code is a little annoying and complicated. There are three main
parts:
1. Walking the package directory and orchestrating builds (lives in
`builder.build.orchestrate)
2. Dealing with downloading and unzipping sources (lives in
`builder.build.download`)
3. Building the wheels (lives in `builder.build.build_wheel`) and the
attendant shellcall infrastructure (lives in
`builder.build.shell_environment`)

## Walking package directories and orchestrating builds

This isn't really that bad; it's a lot of path manipulation. Options
about builds are bundled up and passed around in dataclasses. The intent
is that you can add new stuff in the packages directory and it'll find
it if you put a `build.py` in there. The `build.py` is eval'd and calls
into `builder.build` with a nice interface to specify how to build the
package.

## Downloading and unzipping sources
This is a little more of the same - give us a url, we'll download it,
we'll unpack it (while trying to avoid trivial tarfile path manipulation
attacks). This stuff goes in build/.

## Building the package and handling shells
It's really annoying to deal with the shellcalls. The previous PR's test
file approached it by running the sdk environment prep once, capturing
all the stuff it did, and passing that to calls. This is weird and prone
to failure because passing environment variables to popen for some
reason is super inconsistent.

The slightly more nuclear option, which this PR does instead, is to have
a single long running interactive shell process that you send stuff to
with stdin. now, you can set up the buildroot environment in the
subshell and just send over commands. Works actually pretty well, and
also has some bonuses like "handles aliases correctly" and "you can add
a venv".

In terms of building, we need to handle build-time dependencies, which
we do via a little venv; and we need to set a whole bunch of other magic
environment flags and options that the sdk doesn't do.

Once we do all that, we can build a wheel!


Closes RCORE-209
Closes RCORE-210
Closes RCORE-215