feat(build): Build a PEP503 simple package index from built wheels (#4)
Add a build of a directory structure and html files compatible with a
[PEP503](https://peps.python.org/pep-0503/) simple package index.

When you build the packages, the last step is now to copy everything
from `dist/` to `index/`, with a file structure below index that
fulfills pep503. Index html files contain proper anchor tags for all the
package files with shasum hashes.

Links are all absolute, so the build system gained a flag
`--index-root-url` that lets you specify a url part that all the other
urls will start with. A good way to test it is to use
`--index-root-url:http://localhost:4000`. You can then use the newly
added poe task `serve` to try it out: `poetry poe serve 4000`

When the server is running, you can see that it works by running `pip
download -i http://localhost:4000 --platform=linux_armv7l pandas` and
note that pip fetches the right thing and you get an access log (`poetry
poe serve` uses the http.server builtin so it just puts access logs in
the terminal).

Closes RCORE-214